+++ etl/cite/itp_cite_empresas_contribuyente.py
@@ -60,7 +60,7 @@ class CiteContribuyentePipeline(EasyPipeline):
             'cite_id':                'UInt8',
             'contribuyente_id':       'UInt8',
             'anio':                   'UInt16',
-            'empresas':               'Float32',
+            'empresas':               'UInt32',
             'fecha_actualizacion':    'UInt32'
          }
