+++ etl/cite/itp_cite_empresas_tipo.py
@@ -116,7 +116,7 @@ def run_pipeline(params: dict):
                 'cite_id':               'UInt8',
                 'tipo_cliente_id':       'UInt8',
                 'time':                  'UInt32',
-                'empresas':              'Float32',
+                'empresas':              'UInt32',
                 'fecha_actualizacion':   'UInt32'
             },
             'nullable_list': ['empresas']
